fix adding tags at stack

+++ app/controllers/dispatch_stack_controller.rb
@@ -2,8 +2,8 @@ class DispatchStackController < ApplicationController
   layout "back_end"
 
   IMG_LAST_DIR = 'tmp'
-  IMG_TMP_DIR = Rails.root.join("public","#{IMG_LAST_DIR}") # TODO: bad path
-  GIF = "GIF8"
+  IMG_TMP_DIR = Rails.root.join('app', 'assets', 'images', "#{IMG_LAST_DIR}")
+  GIF = "gif8"
   JPEG = "\xff\xd8\xff\xe0"
   PNG = "\x89\x50\x4e\x47"
 
@@ -15,35 +15,33 @@ class DispatchStackController < ApplicationController
   # TODO: rename actions !!!!
 
   def index
-    if IMG_TMP_DIR.exist? 
-      @img = Dir["#{IMG_TMP_DIR}/*"].find { |e| check_img?(e) }
-      @img = "/#{IMG_LAST_DIR}/#{File.basename(@img)}"
+    if IMG_TMP_DIR.exist? && file = Dir["#{IMG_TMP_DIR}/*.{jpg,jpeg,png,gif}"].first
+      @img = "#{IMG_LAST_DIR}/#{File.basename(file)}" if check_img?(file)
       @tags = Tag.order("name ASC").all
     else 
-      redirect_to dispatch_imgs_path, { alert: "Dir not found" }
+      redirect_to dispatch_imgs_path, { alert: "Dir not found or empty" }
     end
   end
 
   # TODO: !!!! can't add tags
   # TODO: !!!! tags: [tags]
   def create
-    response = {} 
+    response = {}
     file_path = params[:image] ? Rails.root.join("#{IMG_TMP_DIR}", File.basename(params[:image])) : nil
     if params[:button] == "accept"
-      @img = Image.new(image: File.new(file_path))
+      @img = Image.new(image: File.new(file_path), tag_ids: params[:tag_ids])
       response = @img.save_with_response
     end
-    File.delete(file_path)                                          # FIXME: if file creating failed
+    File.delete(file_path) unless response.has_key? :alert
     redirect_to dispatch_stack_index_path, response
   end
-  
+
   private
     def check_img?(file_path)
-      return false if file_path.scan(/.jpg$|.jpeg$|.gif$|.png$/).blank?
-      file = File.open(file_path.to_s,'rb')
+      file = File.open(file_path.to_s)
       data = file.read(9)
-      file.close()
+      file.close
       head = data[0,4].downcase
-      return head == GIF || head == JPEG || head == PNG
+      return head.casecmp(GIF).to_i.zero? || head.casecmp(JPEG).to_i.zero? || head.casecmp(PNG).to_i.zero?
     end
 end
